requirements.txt
Flask~=3.0.2
mysql-connector-python~=8.3.0
pandas~=2.0.3
selenium~=4.19.0
discogs-client==2.3.0
python-dotenv==0.20.0
google-api-python-client==2.49.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1